refactor: extract fixAll() orchestration to orchestrator.ts
- Move FixAllOptions/FixAllResult/FixerExecutionResult types to orchestrator.ts
- Move fixAll() implementation to orchestrator.ts (fixers/index.ts is now pure barrel)
- Fixers/index.ts: 187 lines → 34 lines ( barrel only)
